+++ src/main/java/com/spring/domain/ProductVO.java
@@ -9,7 +9,7 @@ public class ProductVO {
 	private int product;
 	private String p_name;
 	private int interest_rates;
-	private int limit;
+	private long limit;
 	private Date exp_date;
 	private Date repay_date;
